refactor(core): 重构代码规范并更新依赖配置
- 更新 pre-commit 配置中的依赖版本，包括 pre-commit-hooks 从 v4.5.0 到 v5.0.0
- 将代码格式化工具从 black 和 isort 替换为 ruff，提升代码检查效率
- 移除 flake8 并集成 ruff 的 lint 和格式化功能
- 更新 mypy 版本从 v1.8.0 到 v1.14.0
- 在 pre-commit 中替换 check-merge-conflict 和 detect-private-key 为 check-toml 和 check-ast
- 调整多个 __init__.py 文件中的导入顺序和 __all__ 导出列表，保持一致的排序规范
- 在 adaptive_factor_engine 中使用 StrEnum 替代 str 和 Enum 组合
- 更新类型注解语法，使用 | 替代 Union[int, None] 等旧式写法
- 优化代码格式，包括添加缺失的换行符和调整缩进
- 在 AKShare 数据源中更新导入顺序并改进字符串格式化处理

+++ app/data_sources/normalizer.py
@@ -2,10 +2,8 @@
 数据标准化器
 统一多数据源输出格式，确保列名、日期、数值类型、口径标注一致
 """
-from typing import Optional
+
 import pandas as pd
-import numpy as np
-from app.core.logging import logger
 
 
 class DataNormalizer:
@@ -13,18 +11,31 @@ class DataNormalizer:
 
     # 统一输出列定义
     STOCK_DAILY_COLUMNS = [
-        'trade_date', 'open', 'high', 'low', 'close',
-        'pre_close', 'change', 'pct_chg', 'volume', 'amount',
+        "trade_date",
+        "open",
+        "high",
+        "low",
+        "close",
+        "pre_close",
+        "change",
+        "pct_chg",
+        "volume",
+        "amount",
     ]
     INDEX_DAILY_COLUMNS = STOCK_DAILY_COLUMNS
     STOCK_BASIC_COLUMNS = [
-        'ts_code', 'symbol', 'name', 'industry', 'market',
-        'list_date', 'status',
+        "ts_code",
+        "symbol",
+        "name",
+        "industry",
+        "market",
+        "list_date",
+        "status",
     ]
-    TRADING_CALENDAR_COLUMNS = ['trade_date', 'is_open', 'pretrade_date']
+    TRADING_CALENDAR_COLUMNS = ["trade_date", "is_open", "pretrade_date"]
 
     # 成交额为估算值的数据源
-    AMOUNT_ESTIMATED_SOURCES = {'akshare', 'crawler'}
+    AMOUNT_ESTIMATED_SOURCES = {"akshare", "crawler"}
 
     def normalize_stock_daily(self, df: pd.DataFrame, source: str) -> pd.DataFrame:
         """
@@ -40,39 +51,38 @@ class DataNormalizer:
         df = df.copy()
 
         # 1. 列名统一
-        column_map = {'vol': 'volume'}
+        column_map = {"vol": "volume"}
         df.rename(columns=column_map, inplace=True)
 
         # 2. 日期格式统一 → YYYY-MM-DD 字符串
-        df['trade_date'] = self._normalize_date_column(df['trade_date'])
+        df["trade_date"] = self._normalize_date_column(df["trade_date"])
 
         # 3. 数值类型统一
-        numeric_cols = ['open', 'high', 'low', 'close', 'pre_close',
-                        'change', 'pct_chg', 'volume', 'amount']
+        numeric_cols = ["open", "high", "low", "close", "pre_close", "change", "pct_chg", "volume", "amount"]
         for col in numeric_cols:
             if col in df.columns:
-                df[col] = pd.to_numeric(df[col], errors='coerce')
+                df[col] = pd.to_numeric(df[col], errors="coerce")
 
         # 4. 计算 change（如果缺失）
-        if 'change' not in df.columns or df['change'].isna().all():
-            if 'close' in df.columns and 'pre_close' in df.columns:
-                df['change'] = df['close'] - df['pre_close']
+        if "change" not in df.columns or df["change"].isna().all():
+            if "close" in df.columns and "pre_close" in df.columns:
+                df["change"] = df["close"] - df["pre_close"]
 
         # 5. 口径标注
-        df['data_source'] = source
-        df['amount_is_estimated'] = source in self.AMOUNT_ESTIMATED_SOURCES
+        df["data_source"] = source
+        df["amount_is_estimated"] = source in self.AMOUNT_ESTIMATED_SOURCES
 
         # 6. 只保留标准列 + 标注列
         keep_cols = [c for c in self.STOCK_DAILY_COLUMNS if c in df.columns]
-        extra_cols = ['data_source', 'amount_is_estimated']
+        extra_cols = ["data_source", "amount_is_estimated"]
         # 保留 ts_code 如果存在（批量获取时）
-        if 'ts_code' in df.columns:
-            extra_cols.insert(0, 'ts_code')
+        if "ts_code" in df.columns:
+            extra_cols.insert(0, "ts_code")
         df = df[extra_cols + keep_cols]
 
         # 7. 排序
-        if 'trade_date' in df.columns:
-            df = df.sort_values('trade_date').reset_index(drop=True)
+        if "trade_date" in df.columns:
+            df = df.sort_values("trade_date").reset_index(drop=True)
 
         return df
 
@@ -84,38 +94,37 @@ class DataNormalizer:
         df = df.copy()
 
         # 列名统一
-        column_map = {'vol': 'volume'}
+        column_map = {"vol": "volume"}
         df.rename(columns=column_map, inplace=True)
 
         # 日期格式统一
-        df['trade_date'] = self._normalize_date_column(df['trade_date'])
+        df["trade_date"] = self._normalize_date_column(df["trade_date"])
 
         # 数值类型统一
-        numeric_cols = ['open', 'high', 'low', 'close', 'pre_close',
-                        'change', 'pct_chg', 'volume', 'amount']
+        numeric_cols = ["open", "high", "low", "close", "pre_close", "change", "pct_chg", "volume", "amount"]
         for col in numeric_cols:
             if col in df.columns:
-                df[col] = pd.to_numeric(df[col], errors='coerce')
+                df[col] = pd.to_numeric(df[col], errors="coerce")
 
         # 计算 change
-        if 'change' not in df.columns or df['change'].isna().all():
-            if 'close' in df.columns and 'pre_close' in df.columns:
-                df['change'] = df['close'] - df['pre_close']
+        if "change" not in df.columns or df["change"].isna().all():
+            if "close" in df.columns and "pre_close" in df.columns:
+                df["change"] = df["close"] - df["pre_close"]
 
         # 口径标注
-        df['data_source'] = source
-        df['amount_is_estimated'] = source in self.AMOUNT_ESTIMATED_SOURCES
+        df["data_source"] = source
+        df["amount_is_estimated"] = source in self.AMOUNT_ESTIMATED_SOURCES
 
         # 只保留标准列 + 标注列
         keep_cols = [c for c in self.INDEX_DAILY_COLUMNS if c in df.columns]
-        extra_cols = ['data_source', 'amount_is_estimated']
-        if 'index_code' in df.columns:
-            extra_cols.insert(0, 'index_code')
+        extra_cols = ["data_source", "amount_is_estimated"]
+        if "index_code" in df.columns:
+            extra_cols.insert(0, "index_code")
         df = df[extra_cols + keep_cols]
 
         # 排序
-        if 'trade_date' in df.columns:
-            df = df.sort_values('trade_date').reset_index(drop=True)
+        if "trade_date" in df.columns:
+            df = df.sort_values("trade_date").reset_index(drop=True)
 
         return df
 
@@ -127,20 +136,20 @@ class DataNormalizer:
         df = df.copy()
 
         # 列名统一
-        column_map = {'list_status': 'status'}
+        column_map = {"list_status": "status"}
         df.rename(columns=column_map, inplace=True)
 
         # ts_code 格式校验与统一
-        if 'ts_code' in df.columns:
-            df['ts_code'] = df['ts_code'].str.upper().str.strip()
+        if "ts_code" in df.columns:
+            df["ts_code"] = df["ts_code"].str.upper().str.strip()
 
         # industry 缺失填充
-        if 'industry' in df.columns:
-            df['industry'] = df['industry'].fillna('未知')
+        if "industry" in df.columns:
+            df["industry"] = df["industry"].fillna("未知")
 
         # list_date 格式统一
-        if 'list_date' in df.columns:
-            df['list_date'] = self._normalize_date_column(df['list_date'])
+        if "list_date" in df.columns:
+            df["list_date"] = self._normalize_date_column(df["list_date"])
 
         # 只保留标准列
         keep_cols = [c for c in self.STOCK_BASIC_COLUMNS if c in df.columns]
@@ -156,25 +165,25 @@ class DataNormalizer:
         df = df.copy()
 
         # 列名统一
-        column_map = {'cal_date': 'trade_date'}
+        column_map = {"cal_date": "trade_date"}
         df.rename(columns=column_map, inplace=True)
 
         # 日期格式统一
-        df['trade_date'] = self._normalize_date_column(df['trade_date'])
-        if 'pretrade_date' in df.columns:
-            df['pretrade_date'] = self._normalize_date_column(df['pretrade_date'])
+        df["trade_date"] = self._normalize_date_column(df["trade_date"])
+        if "pretrade_date" in df.columns:
+            df["pretrade_date"] = self._normalize_date_column(df["pretrade_date"])
 
         # is_open 类型统一
-        if 'is_open' in df.columns:
-            df['is_open'] = df['is_open'].astype(int)
+        if "is_open" in df.columns:
+            df["is_open"] = df["is_open"].astype(int)
 
         # 只保留标准列
         keep_cols = [c for c in self.TRADING_CALENDAR_COLUMNS if c in df.columns]
         df = df[keep_cols]
 
         # 排序
-        if 'trade_date' in df.columns:
-            df = df.sort_values('trade_date').reset_index(drop=True)
+        if "trade_date" in df.columns:
+            df = df.sort_values("trade_date").reset_index(drop=True)
 
         return df
 
@@ -187,8 +196,8 @@ class DataNormalizer:
 
         # 尝试转为 datetime 再格式化
         try:
-            dt = pd.to_datetime(series, format='mixed', errors='coerce')
-            return dt.dt.strftime('%Y-%m-%d')
+            dt = pd.to_datetime(series, format="mixed", errors="coerce")
+            return dt.dt.strftime("%Y-%m-%d")
         except Exception:
             pass
